<SBRFACCTAX-791><SBRFACCTAX-784> Новый функционал: server-side паджинация в виджете DictionaryPicker
git-svn-id: https://horse.aplana.com/svn/SBRFACCTAX/trunk/taxaccounting@712 34aa7802-8fb0-764d-beb8-9b5b9d595954

+++ gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/widget/dictionarypicker/client/DictionaryDataProvider.java
@@ -61,7 +61,7 @@ public abstract class DictionaryDataProvider<A extends DictionaryAction<T>, T ex
 			@Override
 			public void onReqSuccess(DictionaryResult<T> result) {
 				updateRowData(offset, result.getDictionaryItems());
-				updateRowCount(result.getSize(), true);
+				updateRowCount(result.getSize().intValue(), true);
 			}
 		});
 	}
